online_store_api/requirements.txt
django==3.1.12
djangorestframework==3.11.2
drf-yasg==1.17.1
django-cors-headers==3.2.1
djangorestframework-recursive==0.1.2
django-extentions
